Reintroduce SHA256 and refactor crypto
SHA256 support will enable the future use of other crypto libraries.
For now, both Blake-2b and SHA256 hashes are saved to the meta.conf.
That way if one goes away in the future, .fw files made with this
version will still work. The main possibility being entertained is the
ability to build fwup w/o libsodium. This saves a ~370-450KB shared
library and that may be useful for some small embedded systems.